requirements.txt
asyncio==3.4.3
numpy==1.22.2
Pillow==9.0.1
pyserial==3.5
pyserial-asyncio==0.6
six==1.16.0
wxasync==0.44
wxPython==4.1.1
